Add shortcut extension for creating regexes
Also expose enumeration methods for internal reuse.
